# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "contact": {
      "email": "eti@cisco.io",
      "name": "API Support"
    },
    "description": "This is a Kosha REST service for exposing many teamwork features as REST APIs with better consistency, observability etc",
    "termsOfService": "http://swagger.io/terms/",
    "title": "Teamwork Connector API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.MultiProject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "projects"
        ],
        "description": "List all projects\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/df7d06840ecdd-returns-a-list-of-projects for more parameter options.",
        "summary": "Get all projects"
      },
      "post": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "projects"
        ],
        "description": "Create single project in the system\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/a0c29f3b330bd-create-project for more parameter options.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.SingleProject"
              }
            }
          },
          "description": "Enter project properties"
        },
        "summary": "Create new project"
      }
    },
...
    "/api/v1/projects/{id}": {
      "delete": {
        "parameters": [
          {
            "description": "Enter project id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "projects"
        ],
        "description": "Delete single project based on the project id logged in the system\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/df61dac913b37-delete-project for more parameter options.",
        "summary": "Delete single project"
      },
      "get": {
        "parameters": [
          {
            "description": "Enter project id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.SingleProject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "projects"
        ],
        "description": "List single project based on project ID\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/9daa306fff1d2-returns-a-project for more parameter options.",
        "summary": "Get single project"
      }
    },
    "/api/v1/projects/{id}/risks": {
      "get": {
        "parameters": [
          {
            "description": "Enter project id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.ReturnedRisks"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "risks"
        ],
        "description": "List all risks in the project\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/3f3555d7cd5d0-get-risks-for-a-specific-project for more parameter options.",
        "summary": "Get all risks associated with specified project"
      },
      "post": {
        "parameters": [
          {
            "description": "Enter project id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "risks"
        ],
        "description": "Create a risk for a project in the system\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/f6dfef8ed1b7f-create-a-risk-on-a-project for more parameter options.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.CreateRisk"
              }
            }
          },
          "description": "Enter project risk properties"
        },
        "summary": "Create a risk for a project"
      }
    },
    "/api/v1/projects/{id}/tasklists": {
      "get": {
        "parameters": [
          {
            "description": "Enter project id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.MultiTaskList"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "tasklists"
        ],
        "description": "Lists task lists based on project ID\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/7ee4827082194-get-all-task-lists-for-a-project for more parameter options.",
        "summary": "Retrieve a project's task list"
      },
      "post": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "tasklists"
        ],
        "description": "Create single tasklist in the system\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/36c110772a850-create-task-list for more parameter options.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.NewTaskList"
              }
            }
          },
          "description": "Enter tasklist properties"
        },
        "summary": "Create new tasklist"
      }
    },
...
    "/api/v1/risks/{id}": {
      "delete": {
        "parameters": [
          {
            "description": "Enter risk id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "risks"
        ],
        "description": "Remove a risk\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/a2a40cec66701-deleting-a-risk-on-a-project for more parameter options.",
        "summary": "Delete a risk"
      },
      "put": {
        "parameters": [
          {
            "description": "Enter risk id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "risks"
        ],
        "description": "Add an update to a project's risk\nPlease refer to https://apidocs.teamwork.com/docs/teamwork/a1d50e01411b8-updating-a-risk-on-a-project for more parameter options.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.CreateRisk"
              }
            }
          },
          "description": "Enter project risk properties"
        },
        "summary": "Update a project risk"
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "models.CreateRisk": {
        "properties": {
          "risk": {
            "$ref": "#/components/schemas/models.Risk"
          }
        },
        "type": "object"
      },
...
      "models.Meta": {
        "properties": {
          "limit": {
            "type": "integer"
          },
          "nextCursor": {
            "type": "string"
          },
          "page": {
            "$ref": "#/components/schemas/models.MetaPage"
          },
          "prevCursor": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "models.MultiProject": {
        "properties": {
          "included": {
            "$ref": "#/components/schemas/models.ProjectIncluded"
          },
          "meta": {
            "$ref": "#/components/schemas/models.Meta"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/models.Project"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "models.MultiTaskList": {
        "properties": {
          "STATUS": {
            "type": "string"
          },
          "tasklists": {
            "items": {
              "$ref": "#/components/schemas/models.Tasklist"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "models.NewTaskList": {
        "properties": {
          "applyDefaultToExistingTasks": {
            "type": "boolean"
          },
          "todo-list": {
            "$ref": "#/components/schemas/models.ToDoList"
          }
        },
        "type": "object"
      },
...
      "models.Project": {
        "properties": {
          "tags": {
            "items": {
              "$ref": "#/components/schemas/models.Relationship"
            },
            "type": "array"
          },
          "activePages": {
            "properties": {
              "billing": {
                "type": "boolean"
              },
              "board": {
                "type": "boolean"
              },
              "comments": {
                "type": "boolean"
              },
              "files": {
                "type": "boolean"
              },
              "finance": {
                "type": "boolean"
              },
              "forms": {
                "type": "boolean"
              },
              "gantt": {
                "type": "boolean"
              },
              "links": {
                "type": "boolean"
              },
              "list": {
                "type": "boolean"
              },
              "messages": {
                "type": "boolean"
              },
              "milestones": {
                "type": "boolean"
              },
              "notebooks": {
                "type": "boolean"
              },
              "riskRegister": {
                "type": "boolean"
              },
              "table": {
                "type": "boolean"
              },
              "tasks": {
                "type": "boolean"
              },
              "time": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "announcement": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/models.Relationship"
          },
          "company": {
            "$ref": "#/components/schemas/models.Relationship"
          },
          "createdAt": {
            "type": "string"
          },
          "createdBy": {
            "type": "integer"
          },
          "customFieldValues": {
            "items": {
              "$ref": "#/components/schemas/models.Relationship"
            },
            "type": "array"
          },
          "defaultPrivacy": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "directFileUploadsEnabled": {
            "type": "boolean"
          },
          "endAt": {
            "type": "string"
          },
          "financialBudgetRelationship": {
            "$ref": "#/components/schemas/models.Relationship"
          },
          "harvestTimersEnabled": {
            "type": "boolean"
          },
          "id": {
            "type": "integer"
          },
          "integrations": {
            "properties": {
              "oneDriveBusiness": {
                "$ref": "#/components/schemas/models.IntegrationsMeta"
              },
              "sharepoint": {
                "$ref": "#/components/schemas/models.IntegrationsMeta"
              },
              "xero": {
                "properties": {
                  "baseCurrency": {
                    "type": "string"
                  },
                  "connected": {
                    "type": "boolean"
                  },
                  "countryCode": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "organisation": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            },
            "type": "object"
          },
          "isBillable": {
            "type": "boolean"
          },
          "isOnBoardingProject": {
            "type": "boolean"
          },
          "isProjectAdmin": {
            "type": "boolean"
          },
          "isSampleProject": {
            "type": "boolean"
          },
          "isStarred": {
            "type": "boolean"
          },
          "lastWorkedOn": {
            "type": "string"
          },
          "latestActivity": {
            "$ref": "#/components/schemas/models.Relationship"
          },
          "logo": {
            "type": "string"
          },
          "minMaxAvailableDates": {
            "properties": {
              "deadlinesFound": {
                "type": "boolean"
              },
              "maxEndDate": {
                "type": "string"
              },
              "minStartDate": {
                "type": "string"
              },
              "suggestedEndDate": {
                "type": "string"
              },
              "suggestedStartDate": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "name": {
            "type": "string"
          },
          "notifyCommentIncludeCreator": {
            "type": "boolean"
          },
          "notifyEveryone": {
            "type": "boolean"
          },
          "notifyTaskAssignee": {
            "type": "boolean"
          },
          "overviewStartPage": {
            "type": "string"
          },
          "ownedBy": {
            "type": "integer"
          },
          "portfolioCards": {
            "items": {
              "$ref": "#/components/schemas/models.Relationship"
            },
            "type": "array"
          },
          "privacyEnabled": {
            "type": "boolean"
          },
          "projectOwner": {
            "$ref": "#/components/schemas/models.Relationship"
          },
          "replyByEmailEnabled": {
            "type": "boolean"
          },
          "showAnnouncement": {
            "type": "boolean"
          },
          "skipWeekends": {
            "type": "boolean"
          },
          "startAt": {
            "type": "string"
          },
          "startPage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "subStatus": {
            "type": "string"
          },
          "tasksStartPage": {
            "type": "string"
          },
          "timeBudget": {
            "$ref": "#/components/schemas/models.Relationship"
          },
          "type": {
            "type": "string"
          },
          "update": {
            "$ref": "#/components/schemas/models.Relationship"
          },
          "updatedAt": {
            "type": "string"
          },
          "updatedBy": {
            "type": "integer"
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/models.Relationship"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "models.ProjectIncluded": {
        "properties": {
          "tags": {},
          "activities": {},
          "companies": {},
          "countries": {},
          "customFieldProjects": {},
          "customFields": {},
          "industries": {},
          "portfolioBoards": {},
          "portfolioCards": {},
          "portfolioColumns": {},
          "projectBudgets": {},
          "projectCategories": {},
          "projectTaskStats": {},
          "projectUpdates": {},
          "users": {}
        },
        "type": "object"
      },
...
      "models.ReturnedRisks": {
        "properties": {
          "included": {},
          "meta": {},
          "risks": {
            "items": {
              "$ref": "#/components/schemas/models.Risks"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "models.Risk": {
        "properties": {
          "impactCost": {
            "type": "boolean"
          },
          "impactPerformance": {
            "type": "boolean"
          },
          "impactSchedule": {
            "type": "boolean"
          },
          "impactValue": {
            "type": "string"
          },
          "mitigationPlan": {
            "type": "string"
          },
          "probabilityValue": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "models.Risks": {
        "properties": {
          "canEdit": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "createdBy": {
            "type": "integer"
          },
          "createdByUserId": {
            "type": "integer"
          },
          "createdOn": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "id": {
            "type": "integer"
          },
          "impact": {
            "type": "string"
          },
          "impactCost": {
            "type": "boolean"
          },
          "impactPerformance": {
            "type": "boolean"
          },
          "impactSchedule": {
            "type": "boolean"
          },
          "impactValue": {
            "type": "integer"
          },
          "lastChangedByUserId": {
            "type": "integer"
          },
          "lastChangedOn": {
            "type": "string"
          },
          "mitigationPlan": {
            "type": "string"
          },
          "probability": {
            "type": "string"
          },
          "probabilityValue": {
            "type": "integer"
          },
          "project": {},
          "projectId": {
            "type": "integer"
          },
          "result": {
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "updatedBy": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "models.SingleProject": {
        "properties": {
          "included": {
            "$ref": "#/components/schemas/models.ProjectIncluded"
          },
          "meta": {
            "$ref": "#/components/schemas/models.Meta"
          },
          "project": {
            "$ref": "#/components/schemas/models.Project"
          }
        },
        "type": "object"
      },
...
      "models.Tasklist": {
        "properties": {
          "DLM": {},
          "complete": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "milestone-id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "pinned": {
            "type": "boolean"
          },
          "position": {
            "type": "integer"
          },
          "private": {
            "type": "boolean"
          },
          "projectId": {
            "type": "string"
          },
          "projectName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "uncompleted-count": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    }
  }
}